projects/taiwan_etf_holdings/requirements.txt
aiohttp>=3.9.0
async-timeout>=4.0
pyyaml>=6.0
psycopg2-binary>=2.9
beautifulsoup4>=4.12
lxml>=4.9
python-dotenv>=1.0
apscheduler>=3.10
